+++ connect/runtime/src/test/java/org/apache/kafka/connect/runtime/StateTrackerTest.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.connect.runtime;
 
 import org.apache.kafka.connect.runtime.AbstractStatus.State;
-import org.apache.kafka.connect.util.MockTime;
+import org.apache.kafka.common.utils.MockTime;
 import org.junit.Before;
 import org.junit.Test;
